requirements.txt
Flask
Flask-Cors
Jinja2
itsdangerous
pymongo
python-dotenv